+++ crates/motsu/src/lib.rs
@@ -1,9 +1,9 @@
 #![doc = include_str!("../README.md")]
 #![allow(deprecated)]
-#[cfg(test)]
 extern crate alloc;
 
 mod context;
+mod precompiles;
 pub mod prelude;
 mod revert;
 mod router;
@@ -940,6 +940,19 @@ mod vm_tests {
     }
 
     unsafe impl TopLevelStorage for ChainChecker {}
+
+    #[motsu::test]
+    fn reads_balance(contract: Contract<Erc20>, alice: Address) {
+        // `contract` is already set up
+        let balance = contract.sender(alice).balance_of(Address::ZERO);
+        assert_eq!(U256::ZERO, balance);
+
+        let owner = msg::sender();
+        let one = U256::from(1);
+        contract.sender(alice)._balances.setter(owner).set(one);
+        let balance = contract.sender(alice).balance_of(owner);
+        assert_eq!(one, balance);
+    }
 
     #[motsu::test]
     fn chain_id(contract: Contract<ChainChecker>, alice: Address) {