+++ crates/motsu/src/prelude.rs
@@ -3,5 +3,6 @@ pub use crate::{
     context::{
         Account, Balance, Contract, ContractCall, FromTag, Funding, VMContext,
     },
+    precompiles::deploy_precompiles,
     revert::ResultExt,
 };